+++ ocr/src/main/java/com/ing/software/ocr/OcrObjects/TicketSchemes/TicketScheme.java
@@ -1,0 +1,16 @@
+package com.ing.software.ocr.OcrObjects.TicketSchemes;
+
+import com.ing.software.common.Scored;
+
+import java.math.BigDecimal;
+
+/**
+ *
+ */
+
+public interface TicketScheme {
+
+    Scored<BigDecimal> getBestAmount();
+
+    String toString();
+}